# docs/admins.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ChapRoom API - Admins",
    "description": "Chaproom Backend Restful-API ( Admins Panel )",
    "version": "1.0"
  },
  "host": "localhost:3000",
  "basePath": "/v1",
  "schemes": ["http", "https"],
  "paths": {
...
    "/admins": {
      "post": {
        "tags": ["Admins"],
        "summary": "create new admin",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "authorization",
            "description": "Bearer Token",
            "in": "header",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CreateOrUpdateAdmin" }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "examples": {
              "application/json": {
                "statusCode": 201,
                "data": {
                  "message": "درخواست شما با موفقعیت ثبت شد"
                },
                "error": null
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "examples": {
              "application/json": {
                "statusCode": 422,
                "data": null,
                "error": {
                  "message": "اینپوت پیدا نشد",
                  "details": [{ "phoneNumber": ["شماره تلفن وارد نشده است"] }]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "statusCode": 400,
                "data": null,
                "error": {
                  "message": "این یک خطا است",
                  "details": []
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Admins"],
        "summary": "Get all admins",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "authorization",
            "description": "Bearer Token",
            "in": "header",
            "type": "string",
            "required": true
          },
          {
            "name": "search",
            "description": "search by phoneNumber or name",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "page",
            "description": "set page number",
            "in": "query",
            "example": 0,
            "type": "int",
            "required": false
          },
          {
            "name": "pageSize",
            "description": "set pageSize",
            "example": 25,
            "in": "query",
            "type": "int",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "statusCode": 200,
                "data": {
                  "page": 1,
                  "pageSize": 25,
                  "totalCount": 3,
                  "totalPageLeft": 1,
                  "totalCountLeft": 25,
                  "admins": [
                    {
                      "id": 1,
                      "name": "کسری",
                      "phoneNumber": "09385667274",
                      "active": true,
                      "createdAt": null,
                      "updatedAt": null,
                      "role": {
                        "id": 1,
                        "name": "superAdmin"
                      }
                    },
                    {
                      "id": 2,
                      "name": "کسری",
                      "phoneNumber": "09385667272",
                      "active": true,
                      "createdAt": null,
                      "updatedAt": null,
                      "role": {
                        "id": 1,
                        "name": "superAdmin"
                      }
                    },
                    {
                      "id": 3,
                      "name": "کسری",
                      "phoneNumber": "09385667270",
                      "active": true,
                      "createdAt": null,
                      "updatedAt": null,
                      "role": {
                        "id": 1,
                        "name": "superAdmin"
                      }
                    }
                  ]
                },
                "error": null
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "examples": {
              "application/json": {
                "statusCode": 422,
                "data": null,
                "error": {
                  "message": "اینپوت پیدا نشد",
                  "details": [{ "phoneNumber": ["شماره تلفن وارد نشده است"] }]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "statusCode": 400,
                "data": null,
                "error": {
                  "message": "این یک خطا است",
                  "details": []
                }
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "in": "header"
    }
  },
  "definitions": {
...
    "CreateOrUpdateAdmin": {
      "type": "object",
      "properties": {
        "name": { "type": "string", "example": "کسری میباشد" },
        "phoneNumber": { "type": "string", "example": "09123456789" },
        "password": { "type": "string", "example": "123456789" }
      }
    },
...
  }
}